moved contexts hooks into a folder
